cleanup game_animation.js
